+++ cmd/start.go
@@ -3,6 +3,8 @@ package cmd
 import (
 	"errors"
 	"fmt"
+	schedulepayout "github.com/NodeFactoryIo/vedran/internal/schedule/payout"
+	"net/url"
 	"strconv"
 	"strings"
 
@@ -33,6 +35,12 @@ var (
 	selection      string
 	serverPort     int32
 	publicIP       string
+	// payout related flags
+	payoutPrivateKey           string
+	payoutNumberOfDays         int32
+	payoutTotalReward          string
+	payoutTotalRewardAsFloat64 float64
+	autoPayoutDisabled         bool
 	// logging related flags
 	logLevel string
 	logFile  string
@@ -101,6 +109,19 @@ var startCmd = &cobra.Command{
 		if whitelistArray != nil && whitelistFile != "" {
 			return errors.New("only one flag for setting whitelisted nodes should be set")
 		}
+
+		autoPayoutDisabled = payoutNumberOfDays == 0 && payoutTotalReward == ""
+		if !autoPayoutDisabled {
+			if payoutNumberOfDays <= 0 {
+				return errors.New("invalid payout interval")
+			}
+			rewardAsFloat64, err := strconv.ParseFloat(payoutTotalReward, 64)
+			if err != nil {
+				return errors.New("invalid total reward value")
+			}
+			payoutTotalRewardAsFloat64 = rewardAsFloat64
+		}
+
 		return nil
 	},
 }
@@ -198,6 +219,28 @@ func init() {
 		"20000:30000",
 		"[OPTIONAL] Range of ports which is used to open tunnels")
 
+	startCmd.Flags().StringVar(
+		&payoutPrivateKey,
+		"private-key",
+		"",
+		"[REQUIRED] Loadbalancers wallet private key, used for sending funds on payout",
+	)
+
+	startCmd.Flags().Int32Var(
+		&payoutNumberOfDays,
+		"payout-interval",
+		0,
+		"[OPTIONAL] Payout interval in days, meaning each X days automatic payout will be executed")
+
+	startCmd.Flags().StringVar(
+		&payoutTotalReward,
+		"payout-reward",
+		"",
+		"[OPTIONAL] Total reward pool in Planck",
+	)
+
+	_ = startCmd.MarkFlagRequired("private-key")
+
 	RootCmd.AddCommand(startCmd)
 }
 
@@ -230,6 +273,11 @@ func startCommand(_ *cobra.Command, _ []string) {
 		log.Fatal("Unable to set whitelisted nodes ", err)
 	}
 	log.Debugf("Whitelisting set to: %t", whitelistEnabled)
+
+	if !autoPayoutDisabled {
+		lbUrl, _ := url.Parse("http://" + publicIP + ":" + string(serverPort))
+		schedulepayout.StartScheduledPayout(payoutNumberOfDays, payoutPrivateKey, payoutTotalRewardAsFloat64, lbUrl)
+	}
 
 	tunnel.StartHttpTunnelServer(tunnelServerPort, pPool)
 	loadbalancer.StartLoadBalancerServer(configuration.Configuration{